fewer

+++ src/brokrest/topos/probs.py
@@ -32,4 +32,4 @@ class Window:
 
     def __call__(self, dists: np.ndarray, /) -> np.ndarray:
         in_range = (dists >= self.lower) & (dists <= self.upper)
-        return dists * in_range.float()
+        return dists * in_range.astype(float)
